demo/requirements.txt
beautifulsoup4==4.11.2
boto3==1.34.97
numpy==1.22.3
openai==1.43.0
Pillow==10.4.0
#playwright==1.46.0
streamlit==1.34.0
tqdm
retry
google-generativeai>=0.3.0
selenium>=4.10.0
webdriver-manager>=4.0.0
python-dotenv==1.0.0